# src/main/resources/openapi.yml
openapi: 3.0.0
info:
  title: MetaMoteur API
  description: API du méta-moteur de recherche orienté communauté
  version: 2.0.0
  contact:
    name: MetaMoteur Team
  license:
    name: GPL-2.0
    url: https://www.gnu.org/licenses/old-licenses/gpl-2.0.html

servers:
  - url: http://localhost:8080
    description: Development
  - url: https://api.metamoteur.com
    description: Production

tags:
  - name: Search
    description: Endpoints de recherche publique
  - name: Agent
    description: Communication inter-agents
  - name: Analytics
    description: Statistiques et analytics
  - name: Admin
    description: Administration (authentification requise)

paths:
  /api/search:
    post:
      tags:
        - Search
      summary: Effectuer une recherche
      operationId: search
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchRequest'
      responses:
        '200':
          description: Recherche réussie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '400':
          description: Requête invalide
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Trop de requêtes (rate limit)

  /api/search/redirect:
    get:
      tags:
        - Search
      summary: Redirection avec tracking
      parameters:
        - name: url
          in: query
          required: true
          schema:
            type: string
            format: uri
      responses:
        '302':
          description: Redirection
        '400':
          description: URL invalide

components:
  schemas:
    SearchRequest:
      type: object
      required:
        - query
      properties:
        query:
          type: string
          minLength: 2
          maxLength: 500
          example: "spring boot tutorial"
        maxResults:
          type: integer
          minimum: 1
          maximum: 100
          default: 20
        engine:
          type: string
          enum: [ google, bing, duckduckgo, local ]
          default: google
        useCache:
          type: boolean
          default: true
        contactAgents:
          type: boolean
          default: true

    SearchResponse:
      type: object
      properties:
        query:
          type: string
        results:
          type: array
          items:
            $ref: '#/components/schemas/SearchResultDTO'
        count:
          type: integer
        timestamp:
          type: string
          format: date-time
        fromCache:
          type: boolean
        engine:
          type: string
        metrics:
          $ref: '#/components/schemas/MetricsDTO'

    SearchResultDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        url:
          type: string
          format: uri
        title:
          type: string
        description:
          type: string
        rank:
          type: integer
        clickCount:
          type: integer
        score:
          type: number
          format: double
        source:
          type: string

    MetricsDTO:
      type: object
      properties:
        searchTimeMs:
          type: integer
          format: int64
        agentsContacted:
          type: integer
        agentsResponded:
          type: integer
        localPermutationApplied:
          type: boolean

    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        details:
          type: string
        timestamp:
          type: string
          format: date-time
        path:
          type: string
        status:
          type: integer
        validationErrors:
          type: object
          additionalProperties:
            type: string

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT